requirements.txt
aioredis==2.0.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.25.0
click==8.1.3
databases==0.6.0
fastapi==0.78.0
fastapi-asyncpg==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
inflect==5.6.0
Mako==1.2.0
MarkupSafe==2.1.1
pydantic==1.9.1
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.37
starlette==0.19.1
typer==0.4.1
typing-extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
